Improve workshop.ripley in order to reset the database properly
Note that it generates a lot of "Orphan detected" warnings. It needs
investigate but it should work even with this "problem".

A feature is still missing. A maintenance mode must be implemented in
order to be sure that it's not possible to do something in the app while
a ripley is running.