Skip after-create if a second .created event is received
It's possible to receive two .created event when more than one create is
called shortly. When the first create has just sent the event, maybe
according to the node event loop, the second create is sending this
event too.

In this case, the second event must skip the after-create hook.