Try to fix potential deadlock like with entity load-graph
When next.parallel is used, the result is always an array. It seems that
this array is always defined but it can be just empty (length === 0).
Instead of just testing if defined, it's necessary to check the length.

The deadlock appears when two `yield next.sync()` are called
consecutively. It's the result of multiple hours of debugging and
traces. But most of time is works (the generator continue). It's a bit
difficult to tell why the generator continue or not. But it's sure
that's wrong to call two times next.sync().
I hope that this commit will fix the deadlock.