List/search mode: impl. facet count reloading when index change
When data is indexed, or unindexed, i update aggregations and total
count, and refetch the list.

Range filter is now detecting the "all" mode, for updating from/to to limits
when aggegration detect new min/max values.

This commit refactor the logic handlers for mutating precisly filter
values (avoid loose of data, when updating filters)
and using mutableReducer when adapted.