Entity-Indexer: add new goblin in charge of indexing
When an entity is indexed via the rehydrate process,
an event is emitted for the entity-indexer.

The entity indexer, take care of collecting indexation
requests, bulking requests to one call every seconds.

The called elasticsearch bulk quest is itself limited to 50 in parallel.

the resulting indexation, can emit usefull events
for search lists, theses events are limited to 1/sec/type
(avoiding overload)