rework details service
set-entity was leaking a lot because we kill and create too closely.

this rework keep a cache of 10 entries with fully async loaded workitems graph.
when the cache is full, we wait for killing old entries and we restart a fresh cache.

i also removed debouncing of set-detail events when navigating in the hinter.

the detail service can handle multiple set, the UX is responsive.