common.registerQuest is now using sysCall
This ensure that the registered quest will do a custom
syscall before running.

This prevent workitems and entities quests from crossfeeding.

ex:
when a registered quest do a create using the system feed,
the only feed running this quest is generally a desktop feed.

with the new leaks preventions rules in warehouse, we must ensure
that the caller service is in system feed before creating other services
in system feed.