Fix sync copy with makeDeepCopy
the storage function provided to makeDeepCopy
was running in fire&forget mode. Racing along the creation.

Conducting to finish a "copy" create before persistance was done.

I rewritted makeDeepCopy for yielding on the store func,
I also moved arguments, because with watt, providing a func as last arguments
cause malfunction.