Create the hinter in readonly mode too
It fixes a bug when a workitem is created in 'readonly' and a new workitem
will use the first workitem in edit mode. Before the commit, the
setDetail was broken because the hinters were not created by the first
create (readonly workitem). Now the hinters are created in both modes;
only the backend mode skip the hinters.

The new 'noHinters' flag forwarded by the setEntity quest of the detail
service prevents an infinite loop when an entity loads sub-workitems for
the same type of entities.